s: leak callbacks and move to flex module
This is the best flexible API I can come up with for now but it is not
flexible enough for me to provide an ergonomic API on top of it. So I
stick with the flex/ergo organization in order to isolate all unsafe
code to a single library crate.