feat(mdb): Improve Error
There are two use-cases of Error:
 - Owned: When mdb-sys functions fail they return an owned Error.
 - Borrowed: The callbacks passed to mdb-sys functions accept a
   reference to an Error from mdb-sys.

There are now two error types: Error (public to consumer) and
BorrowedError (public to the crate). The owned error returned by
mdb-sys is transparently an Error.

The references to Error are wrapped by BorrowedError within the crate
and exposed as &Error to consumers. This is possible because both types
wrap a pointer to mdb_sys::mdb_error so BorrowedError can be coerced
into an Error.