feat: Add `fleet-manager` program (#111)
This program makes it easier to develop for a variety of hardware or
to develop apps that are distributed over multiple devices. This may
be a niche use, but since I find it useful I may as well share it and
see if it helps anyone else. If it doesn't I may well end up removing
it again.

I wrote most of the program a long time ago, so motivation will be
sparse.  

`crates/fleet-manager/src/main.rs`:
- Logging is set up differently than in other programs because I *hope*
  to use tracing spans to differentiate between logs from different
  threads.
- Don't read `AXIS_DEVICE_IP` because users will likely want to change 
  it between invocations and using the same IP twice is likely a
  mistake.
- Other parameters are not read from environment variables to be
  consistent, but since these are conceivably the same across all
  devices that are adopted this may change in the future. In particular
  the password would be good to read from env since providing passwords
  as arguments is a bad practice.