chore: Use workspace dependencies consistently
After #99 started checking that app builds are reproducible I find
myself spending a lot of time making sure I don't accidentally upgrade
a dependency that is used directly or indirectly by an app while
working on one of the tools. Hopefully, by delegating the task of
specifying a version to the workspace manifest, the number of
accidental version bumps will decrease.