Don't install the `devcontainer` program by default (#128)
The method I used for installing `npm` was flaky and a common source of
false positives in CI. Since it is not needed in CI, and it is needed
locally only when not using the recommended dev container based
environment, not installing it by default seems like a cost effective
remedy.

The steps to install it are still kept to make it easier to reproduce
my development environment. It expects that a `venv` is already
`source`d but does not eagerly verify this. Since it doesn't look like
it has any bad side effects and the script will be used only rarely and
is not easily discoverable this feels like a worthwhile trade-off.