feat: Add default username and password (#130)
The values chosen are the same as was already used in the `Makefile`.
If any default values should be set, these are probably the best. But
it is possible that some users are better served by having no default
values, and if that is the case we should try to make a more thorough
analysis of the net benefit of providing these defaults.