chore: forbid unsafe code
Writing unsafe code is difficult and should generally be avoided. When
using unsafe code is necessary, it is ideal if it can be delegated to
public crates that can be tested and audited by multiple developers.

This commit consistently forbids unsafe code in crates where it is not
necessary to serve as a reminder that it should probably be avoided
also in the future and that the intention of this project is to provide
safe abstractions that allow application authors to avoid unsafe code
in their applications.

`apps-aarch64.checksum`:
- All checksums change, presumably because the line numbers have all
  changed.