s: Give user full control of staging dir
Since we don't do any incremental compilation and the future should have no staging dir at all.

Bound the lifetime of the builder and return a string instead of a path to make it clear to the user that the result is dependent on the existence of the path they passed in.

A better interface may be to make the staging dir entirely an implementation detail and instead return a named temporary file and a final eap name.