feat!(cargo-acap-build): Make the release profile the default
Since `cargo-acap-sdk` is the preferred tool for local development it
should be relatively rare for people to use `cargo-acap-build` locally.
As such it makes sense to optimize it for the continuous integration
where I expect it will be used relatively frequently.

This decision rests on the assumption that there are few cases when a
developer would be better server by `cargo-acap-build` for local
development compared to `cargo-acap-sdk build`.