feat: Include commit id in version reported by CLI
There are essentially two ways to accomplish this:
1. Using a build script that writes a file or sets an environment
   variable.
2. Using a procedural macro.

`clap` recommends `shadow-rs` which uses the first method, but they
don't motivate why or compare it with any other approaches.

`rustup` switched from the first to the second, noting that the second
method is superior because the result does not get cached. I suspect
this could be subjective, but at least this constitutes one reputable
project having considered both approaches and decided to go with the
second.

Unlike `rustup` I write my own procedural macro because I'm weary of
`git-testament` extracting more information than I need, in particular
information that is hard to reproduce, and because I could not get it
to work together with `clap`s derive feature.