Run FFI patterns under Miri
The new tests are included during normal test runs because to make
sure they are correct also outside of miri; I accidentally introduced
a concurrency bug by factoring out the code that uses a static atomic
int and storing 0 to reset the count between runs. But outside of miri
the tests run in parallel and this was insufficient.

I have tested a variety of ways to make the code easier to read. Most
failed under miri, but I plan to introduce the one that works in a
later commit.

wip: start mocking axevent bindings for Miri

Compared to working with a model

Pros:
- Testing the actual API

Same:
- Not testing the real implementation

Cons:
- A lot of work
- Harder to see the patterns and convince myself that they are the same in other crates
  - Alternatively: a lot of work to do this for other crates too.
- Probably cannot use real unit tests in miri, and maybe vice versa
  - Alternatively: a lot of work to make a faithful mock.
  - If the tests are disjoint then what is ensuring the API is equivalent?

wip: replace bindings with mock

f: working towards a mock that makes send_event pass

Many functions are never used and I removed them to make it easier to
navigate the code that does matter. All the element item methods make
sense since I intentionally left that out of the bindings, but I was
surprised that ax_event_key_value_set_remove_key is not removed. There
may be more unused functions still.

The next problem is that we use glib functions to convert between
system time and DateTime. Since these come from an so file, miri is
unhappy.

I'm also a bit unsure about how to handle pointers on the C side.
Options include:
* Prefer *_ T
* Prefer T
* Prefer Box<T>

f: mock glib::DateTime

Still struggling with

f: mock more of glib

Trying to minimize diff in safe abstractions.
To do so I find myself having to mock a lot of glib.

Should try the third approach soon and see how that works out.

Change-Id: I3c31dc378fa5e3e68cb80a36ec07f58150c75879