Run FFI patterns under Miri
The new tests are included during normal test runs because to make
sure they are correct also outside of miri; I accidentally introduced
a concurrency bug by factoring out the code that uses a static atomic
int and storing 0 to reset the count between runs. But outside of miri
the tests run in parallel and this was insufficient.

I have tested a variety of ways to make the code easier to read. Most
failed under miri, but I plan to introduce the one that works in a
later commit.