fix(acap-build): Add `--build=no-build` option (#142)
This is needed to build the apps in `acap-native-sdk-examples`. With
this change all the examples except for one can be reproduced:
`openssl_curl_example` cannot be reproduced but that is probably only
because `libcrypto.so` is not reproducible. Since building this tools is
not central to this project and the tests demonstrating this are a bit
messy I choose to not track them at this time.