f: Align trampoline with axevent
This helps readers familiar with one understand and audit the other.
It also reduces the risk that one of the patterns is unsound but we
don't discover it because the patterns compete for test coverage.

Maybe someone who knows unsafe Rust well could say that they are or are
not equivalent, but I am still not 100% confident about how casting and
dereferencing pointers works.