fix(axevent): Free callbacks on error (#144)
While reviewing #115 I realized that the callbacks in `axevent` are not
dropped if the C function returns an error. This commit fixes that.

It also adopts the `Deferred` implementation from the aforementioned
PR; I'm weary of moving more unsafe logic into `Deferred` because I
find it is difficult to reason about unsafe abstractions but one of the
implementations should change for consistency and the conciseness of
the new implementation is tempting.