r: Add lifetime to KeyValueSet
This lets us return a KeyValueSet from Event without storing an intermediate value.

If this is sound, it's an intriguing option to make it easier to keep track of lifetimes.
That said I'm not sure I like the interface since all functions that deal with KVS now need to be generic over the lifetime.