s: Move script into bin/
Part of the idea with this map is that it can be added to the PATH making the scripts therein easy to access.
Since this script will probably not be executed directly, this may not be a perfect fit.

In any case it helps keep the repo root a little bit tidy.

Change-Id: I7b69f1c76dc62957b0f346094ef12fb83a591852