fix(licensekey)!: Return `CStringPtr` instead of `CString` (#151)
The previous implementation was also sound, but used an unnecessary
`clone`. Though a minor cost since the string is unlikely to be massive
and the `clone` likely to be rare, the fix is straight forward and
relies on a recurring pattern so the increased complexity is low. The
main downside is thus that the user has to do slightly more work to get
an owned `CString` or `String` if that is something they need.

`crates/licensekey/src/flex.rs`:
- Use `CStringPtr` instead of `GString` to avoid adding `glib` to the
  public API.
- Split unsafe blocks to make it easier writing targeted safety
  comments.