refactor(mdb): Split into more modules
The layout is roughly the same as the C++ bindings. Since the languages
are similar, the bindings will likely be similar too, and having a
comparable organization could make the bindings easier to compare.

The main difference that I foresee is that in Rust the module and the
crate are important boundaries when working with unsafe code, a concept
that does not translate to contemporary C++.