s: Implement most of channel
I'm considering making the API similar to the C and C++ APIs at first
to make it easier to verify that they match and, only once that is done,
refactor the code to make it idiomatic Rust. To that end I am leaving
lots of `TODO`s for myself. I am however changing the docstrings
eagerly, but I may go back on this if the upstream authors are
interested in seeing a diff.

Now I'm stuck on the on_done callback for publish; somehow I must free
the user_data from the callback, or when dropping the channel. It looks
like the C++ bindings have implemented this but C++ is difficult for me
to parse.