Box receiver part of channel
This causes the entire struct to be `Unpin` which means it does not need
to be pinned when used. One downside of this approach is that we need to
heap allocate the receiver meaning this crate will not work in
`[no-std]` environments. However, given that we target Axis network
cameras this is not a use case we need to support.

By boxing this future the stream trait implementation is greatly
simplified and still provides good ergonomics to the api consumer.