Provide a futures_lite::Stream implementation for event subscription (#174)
Provides an api that can be composed together with different asynchronous APIs which is useful in an I/O bound ACAP application since it allows different tasks to run concurrently in one thread. Intended to be used together with https://gtk-rs.org/gtk-rs-core/git/docs/glib/struct.MainContext.html#method.spawn and the sister method `spawn_local` which allows running Future's on the GMainLoop. The GMainLoop needs to be running for the axevent wrapper to work as intended. Additionally, Glib can be linked dynamically through (and already is through for example the axevent crate) meaning an entire separate async runtime such as `tokio` can be skipped which results in smaller binaries.
A futures_lite::Stream implementation allows an API that is ergonomic to use without having to spawn a background thread.