chore(deps): bump glib and glib-sys to 0.20.9 (#180)
* chore(deps): bump glib and glib-sys to 0.20.9

New version includes fix for an issue with VariantStrIter that causes crashes.

`Cargo.lock`:
Since glib was updated some of it's dependencies were also bumped in the
lockfile.

`apps-aarch64.checksum`:
Updated checksums are to be expected since most examples depend on glib
either directly or indirectly. Apps that do not depend on glib are
unchanged.

`apps-aarch64.filesize`:
File size changes are to be expected since most examples depend on glib
either directly or indirectly. Apps that do not depend on glib are
unchanged.