fix(inspect_app): Print environment variables individually
Collecting the environment variables into a `HashMap` made the order in
which they were printed unstable, making it unnecessarily difficult to
compare the output.

It also obscured the order of environment variables, which seems to be
stable. I don't know when this would ever have been a problem, but
considering that the purpose of the app is to describe the environment
it runs in accurately this seems like a downside.

Another small problem was that the log spanned multiple lines which can
be annoying when using simple text processing tools such as `sort` to
analyze the logs. With such small logs this was not a big issue to
begin with and now that the order is stable it is even less of an
issue.

Printing the environment variables individually solves all of these
issues.

The only part of the output that is not stable now are the SSH
variables because they contain a port number that changes every time.