feat(acap-logging): Make env_logger optional and remove syslog
Using the app profile, the size of hello_world is:
- 387K (before)
- 323K (after, with default features)
- 286K (after, without default features)

For reference the size of a no-op application compiled with the same
settings is 263K. These are all numbers for aarch64-unknown-linux-gnu.

For some reason `du` and `stat` reports that the size is no smaller
with the default features disabled, but `size` shows that there is
a difference and when run the expected difference in function is also
observed.