feat(acap-logging): Make env_logger optional and remove syslog (#185)
The reason is to make the cost in terms of storage smaller because this
is a utility that most applications would use and the cost adds up if
many applications are installed at the same time.

Using the app profile, the size of hello_world is:
- 387K (before)
- 323K (after, with default features. `size` total is 323,436)
- 323K (after, without default features. `size` total is 286,169)

For reference the size of a no-op application compiled with the same
settings is 263K. These are all numbers for aarch64-unknown-linux-gnu.

Why the size as reported by `du` and `wc` does not decrease when I
expect it to is a mystery to me, but it's not a rabbit hole I want to
explore further at this time.

One bonus benefit of this change is that the device information will
no longer be duplicated in the syslog.
I.e. it will say:

> May 02 17:17:38 abc hello_world[20894]: Logging initialized

instead of:

> May 02 17:17:38 abc hello_world[20894]: abc hello_world[20894]: Logging initialized


---

`apps/*/Cargo.toml`:
- Enable `acap-logging/default` features to maintain the ability to
  run the apps interactively.
- Propagate the `acap-logging/default` feature to make it possible to
  compile `hello_world` with tty support, as one would during
  development, and without it, as one would when deploying the app to 
  production. This serves to document one way in which this can be
  achieved and also to make it easier to test that it works as expected
  (see `docs/manual-tests.md`).

`apps/reverse_proxy/Cargo.toml`,
`crates/axparameter/Cargo.toml`:
- Remove `syslog` and `log` because they were not used and `syslog`
  should not be used.

`crates/acap-logging/Cargo.toml`:
- Bump the version to make these improvement instantly releasable
  because they are meaningful and it is unclear when the next
  improvements to the crate would be made.

`docs/manual-tests.md`:
- Added to start documenting some of the use cases that the apps and
  tools have been designed for because it can be difficult to remember
  and new contributors don't even have the chance to remember. I view
  this as an experiment and may change the approach to documentation or
  go back to relying on git logs.