chore: Link to article "Managing the size of Rust binaries" (#187)
The large size of Rust binaries is a well known problem and search
engines will readily return many results on the topic, not least the
excellent min-sized-rust. However, none of these results focus
specifically on use cases that are similar to ACAP apps. The linked
project is an effort to collect and organize information with
developers of ACAP-like applications in mind.