s: Provide separate ports for http and https
This is done primarily to make is easier to set the environment
variables programmatically, because then both may be available but it
is not known, or tedious to specify, which the user intends to use.

A positive side effect of this decision is that tools will be able to
use both HTTP and HTTPS, and that it scales to other services such as
SSH. The environment variable for the SSH port would be called
`AXIS_DEVICE_SSH_PORT` and a pattern starts to emerge; the name follows
a pattern that looks something like `AXIS_DEVICE_{service}_{property}`.