fix(*): Treat MacOS on ARM as host, not target (#204)
This improves the experience of working on the project on an Apple
computer with apple silicon. Getting the bindings to compile is still
difficult, but this allows running `cargo check` on host. It may also
be necessary to enable working on the other packages on host, but I
have not explicitly tested that.

The implementation to this commit does not solve the root problem;
if someone is trying to develop of an ARM based computer it is
unlikely that we will be able to use the `target_*` keys to classify
that as host. But in terms of classification it is a strict improvement
over status quo.