feat(acap-ssh-utils)!: Add option to override the SSH port. (#199)
This is needed to use the run and test commands from `cargo-acap-sdk`
with the Virtual Loan Tool, which uses port mapping to make multiple
cameras available at one (or possibly a small number) of IP addresses.

Ideally, we would only bump them when we do a release, but we don't do
proper releases, and I still want to be able to ask users what version
they are on when helping them to troubleshoot. The need for aggressive
version bumping will hopefully be eliminated with the merging of #203.