fix(acap-ssh-utils): Disable strict host key checking (#202)
If the server had not been added to known hosts, the program would exit
with a cryptic "Error: Broken pipe (os error 32)".
This is the second of two such errors I often encounter when setting up
a device manually. The other luckily comes with a bit more context:

> ssh: connect to host ... port ...: Connection refused
> Error: Broken pipe (os error 32)

That error can be encountered for once of three reasons to my knowledge:
- SSH has not been enabled
- Developer mode has not been configured, and the user consequently
  does not exist.
- The password of the SSH user has not been set up.

But that is another issue for another time.

I'm not crazy about the security implications of this change, but;
- The tools are usually used in a somewhat safe environment.
  This is why it's OK to fall back on HTTP.
- The information in an EAP typically isn't highly sensitive.
  This is not a guarantee, and I would like to find a better way.