fix(axevent): Handle null when reading from key value sets (#207)
Previously `get_string` would panic if the value had been set to null.
I think the reason this case wasn't considered originally is that none
of the other getter methods make it possible to explicitly detect null
values.

This implementation tests up to two initial values to determine with
certainty if the value is set to `null` because it makes the interface
consistent and idiomatic.

The main alternative considered was mapping both `null` and the initial
values to the same value, `None`. This would have the benefit of a
minimal runtime overhead, thus following the principle that these ACAP
Native API wrappers should minimize runtime overhead and maximize
flexibility. But the overhead is probably relatively low compared to
sending and receiving an event, and we agreed the improved API is worth
the cost in this case.

Another alternative considered was passing the initial value through to
the rust caller, allowing them to retry if possible. This would
probably be the truest to the min-overhead, max-flexibility principle,
but this also doesn't seem worth it.

---

`crates/axevent/src/flex.rs`:
- Use `T:MIN` as the first sentinel value because it is probably rare
  making it unlikely that we have to go to the second iteration.
  A somewhat random value close to `T:MIN` or `T:MAX` would probably be
  rarer still, but it also wouldn't register as suspicious in case
  someone has to troubleshoot the code.
- Add a test documenting the fact that the underlying key value set can
  return a value other than `GFALSE` or `GTRUE` when storing booleans.